feat: centralize RPO navigation management
