+++ rpo-training/index.html
@@ -54,7 +54,7 @@
                     </div>
                 </div>
                 <div class="mt-8 grid md:grid-cols-2 gap-6">
-                    <div class="module-card" onclick="navigateTo('module-1')">
+                    <div class="module-card" onclick="NavigationManager.navigate('module-1')">
                         <p class="text-sm font-semibold text-blue-600">MODULE 1</p>
                         <h3 class="google-sans text-xl font-bold mt-2">Prompting &amp; Writing — The C.R.E.A.T.E.
                             Framework</h3>
@@ -62,7 +62,7 @@
                             (Character, Request, Examples, Additions, Type, Extras) for crafting precise, repeatable,
                             high-quality prompts.</p>
                     </div>
-                    <div class="module-card" onclick="navigateTo('module-2')">
+                    <div class="module-card" onclick="NavigationManager.navigate('module-2')">
                         <p class="text-sm font-semibold text-blue-600">MODULE 2</p>
                         <h3 class="google-sans text-xl font-bold mt-2">Sourcing & Research</h3>
                         <p class="mt-2 text-gray-600">Generate complex Boolean strings and perform rapid market
@@ -87,18 +87,18 @@
                     </div>
                 </div>
                 <div class="mt-8 grid md:grid-cols-3 gap-6">
-                    <div class="module-card" onclick="navigateTo('module-3')">
+                    <div class="module-card" onclick="NavigationManager.navigate('module-3')">
                         <p class="text-sm font-semibold text-green-600">MODULE 3</p>
                         <h3 class="google-sans text-xl font-bold mt-2">Data & Knowledge</h3>
                         <p class="mt-2 text-gray-600">Analyze data in Sheets and build a "CV Bot" to search documents.
                         </p>
                     </div>
-                    <div class="module-card" onclick="navigateTo('module-4')">
+                    <div class="module-card" onclick="NavigationManager.navigate('module-4')">
                         <p class="text-sm font-semibold text-green-600">MODULE 4</p>
                         <h3 class="google-sans text-xl font-bold mt-2">Automation</h3>
                         <p class="mt-2 text-gray-600">Build hands-free workflows that connect Gmail, AI, and Sheets.</p>
                     </div>
-                    <div class="module-card" onclick="navigateTo('module-5')">
+                    <div class="module-card" onclick="NavigationManager.navigate('module-5')">
                         <p class="text-sm font-semibold text-green-600">MODULE 5</p>
                         <h3 class="google-sans text-xl font-bold mt-2">Train the Trainer</h3>
                         <p class="mt-2 text-gray-600">Learn to mentor colleagues and present your projects to
@@ -123,13 +123,13 @@
                     </div>
                 </div>
                 <div class="mt-8 grid md:grid-cols-2 gap-6">
-                    <div class="module-card" onclick="navigateTo('module-6')">
+                    <div class="module-card" onclick="NavigationManager.navigate('module-6')">
                         <p class="text-sm font-semibold text-purple-600">MODULE 6</p>
                         <h3 class="google-sans text-xl font-bold mt-2">Strategy & Governance</h3>
                         <p class="mt-2 text-gray-600">Develop an RPO-specific AI roadmap and prioritize high-impact
                             projects.</p>
                     </div>
-                    <div class="module-card" onclick="navigateTo('module-7')">
+                    <div class="module-card" onclick="NavigationManager.navigate('module-7')">
                         <p class="text-sm font-semibold text-purple-600">MODULE 7</p>
                         <h3 class="google-sans text-xl font-bold mt-2">Measuring Impact</h3>
                         <p class="mt-2 text-gray-600">Define success metrics and build a compelling business case for AI
