+++ rpo-training/sessions/1-2.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
 
 <!-- Content from google_site_training_1_2 (updated to C.R.E.A.T.E.) -->
 <div class="content-section text-center">
