+++ rpo-training/sessions/2-1.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_2_1 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 2.1: AI for Advanced Sourcing</h2><p class="mt-4 text-lg text-gray-600">Our Goal: To use AI to move beyond obvious keywords and become true market intelligence experts. We will find better candidates, faster.</p></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">Part A: Boolean & Beyond</h3><p class="text-lg text-gray-600">Stop Brainstorming Keywords. Start Generating Them.</p></div><div class="mt-6 bg-white border rounded-lg p-6"><p class="mb-4">The old way of writing a Boolean string is slow. You try to think of every possible synonym and job title. The new way is to tell the AI what you need and let it do the heavy lifting.</p><h4 class="google-sans text-xl font-medium mb-2">The Evolution of a Search Prompt:</h4><ol class="list-decimal list-inside space-y-2"><li><strong>Start with a simple idea:</strong> "Software Engineer with Python."</li><li><strong>Ask AI to expand it:</strong> "Give me a list of 20 alternative job titles and 30 related technical skills."</li><li><strong>Ask AI to build it:</strong> "Using these titles and skills, build me a complex Boolean search string for LinkedIn."</li></ol><details class="bg-gray-50 border rounded-lg cursor-pointer mt-6"><summary class="google-sans p-4 font-medium">View the "Boolean Builder" Prompt Example</summary><div class="p-4 border-t prompt-box"><strong>[ROLE]</strong> You are an expert tech sourcer with deep knowledge of LinkedIn Recruiter's search syntax.<br><br><strong>[CONTEXT]</strong> I am looking for a "Senior Software Engineer" with experience in Python, AWS, and database management. I need to create a comprehensive search string to find candidates in the Budapest area.<br><br><strong>[TASK]</strong> First, brainstorm a list of 15 alternative job titles and 20 related technical skills for this role. Then, using all of those terms, build me a complex, multi-layered Boolean search string that is optimized for LinkedIn Recruiter.<br><br><strong>[FORMAT]</strong> Provide the brainstormed lists first, then provide the final Boolean string in a code block.</div></details></div></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">Part B: AI as Your Research Assistant</h3><p class="text-lg text-gray-600">Answer Questions in Minutes, Not Hours.</p></div><div class="mt-6 bg-white border rounded-lg p-6"><p class="mb-4">Sourcing isn't just about finding people; it's about understanding the market. AI can be your personal research analyst. Here are three key research prompts:</p><div class="space-y-4"><div class="p-4 bg-gray-50 rounded-lg"><h4 class="google-sans font-medium">1. Market Mapping</h4><p class="text-gray-600">"List the top 10 biotech companies specializing in gene therapy in the Boston area."</p></div><div class="p-4 bg-gray-50 rounded-lg"><h4 class="google-sans font-medium">2. Competitor Analysis</h4><p class="text-gray-600">"Who are the main competitors for Company X, and what are their typical software engineering job titles?"</p></div><div class="p-4 bg-gray-50 rounded-lg"><h4 class="google-sans font-medium">3. Rapid CV Analysis</h4><p class="text-gray-600">"Summarize this CV and create a table comparing the candidate's experience against these 5 key requirements from the job description."</p></div></div><details class="bg-gray-50 border rounded-lg cursor-pointer mt-6"><summary class="google-sans p-4 font-medium">View the CV Summarization Prompt Example</summary><div class="p-4 border-t prompt-box"><strong>[ROLE]</strong> You are a senior RPO recruiter tasked with quickly screening candidates against a job description.<br><br><strong>[CONTEXT]</strong> I have a candidate's CV and the job description for the role they applied for. I need a fast, objective summary of how well they match the key requirements.<br><br><strong>[TASK]</strong> Read the CV and the Job Description provided below. Create a table with three columns: "Key Requirement," "Evidence from CV," and "Match Level (High/Medium/Low)."<br><br><strong>[FORMAT]</strong> The output should be ONLY the table.<br><br>---<br>[PASTE JOB DESCRIPTION HERE]<br>---<br>[PASTE CANDIDATE CV HERE]<br>---</div></details></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Lab Activities</h3><div class="grid md:grid-cols-2 gap-8"><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Lab 1: The Boolean Builder (20 Mins)</h4><p><strong>Your Goal:</strong> Create a comprehensive Boolean string for a niche, hard-to-fill role like "Clinical Research Associate with Oncology experience."</p><ul class="list-disc list-inside mt-4 space-y-1"><li>Work in your breakout groups.</li><li>Use the prompt structure to build your own multi-layered prompt.</li><li>Run the prompt in Gemini.</li><li class="font-semibold">Crucially: Test the resulting Boolean string in a real LinkedIn search!</li></ul></div><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Lab 2: The 5-Minute Market Map (20 Mins)</h4><p><strong>Your Goal:</strong> Quickly gather actionable intelligence on a specific market to inform your sourcing strategy.</p><ul class="list-disc list-inside mt-4 space-y-1"><li>Choose a role and a city.</li><li>Use research prompts to produce a one-page "Market Snapshot."</li><li class="font-semibold">Your snapshot must include: top 5 target companies, 3-5 common job titles, and a summary of one sample profile.</li></ul></div></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Key Takeaways & Next Steps</h3><p class="max-w-2xl mx-auto mb-4">Your new superpower is letting AI do the heavy lifting on Search Logic, Market Research, and Candidate Analysis. The golden rule is to use AI to automate the research so you can spend more time on the human parts of recruiting: strategy and engagement.</p><div class="bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 text-left inline-block"><p><strong class="font-bold">Next Session:</strong> Session 2.2 - The Randstad AI Toolkit: Gemini & Beyond.</p></div></div>
 <div id="footer-placeholder"></div>
