+++ rpo-training/sessions/2-2.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_2_2 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 2.2: The Randstad AI Toolkit</h2><p class="mt-4 text-lg text-gray-600">Our Goal: To eliminate "tool confusion" with a simple decision-making framework that prioritizes security, integration, and efficiency.</p></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">The Decision Framework</h3><p class="text-lg text-gray-600">A Simple 3-Step Process for Choosing the Right Tool</p></div><div class="mt-8 grid md:grid-cols-3 gap-6"><div class="step-box"><div class="flex items-center mb-3"><span class="google-sans bg-blue-500 text-white rounded-full h-8 w-8 flex items-center justify-center text-lg font-bold">1</span><h4 class="google-sans text-xl font-medium ml-3">Check for Sensitive Data</h4></div><p>Is there ANY candidate, client, or internal data involved? If YES, you MUST use our secure, internal tool: Gemini.</p></div><div class="step-box"><div class="flex items-center mb-3"><span class="google-sans bg-blue-500 text-white rounded-full h-8 w-8 flex items-center justify-center text-lg font-bold">2</span><h4 class="google-sans text-xl font-medium ml-3">Check for Integration</h4></div><p>Does the task need to connect with Docs, Gmail, or Sheets? If YES, Gemini for Google Workspace is the most efficient choice.</p></div><div class="step-box"><div class="flex items-center mb-3"><span class="google-sans bg-blue-500 text-white rounded-full h-8 w-8 flex items-center justify-center text-lg font-bold">3</span><h4 class="google-sans text-xl font-medium ml-3">General Creative Tasks</h4></div><p>If you answered NO to the first two questions, a public tool like ChatGPT is an acceptable option for non-sensitive, creative brainstorming.</p></div></div></div><div class="content-section grid md:grid-cols-2 gap-8 items-center"><div><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Gemini's Strengths</h3><div class="space-y-4"><div><h4 class="google-sans text-xl font-medium">1. Data Security (The #1 Priority)</h4><p class="text-gray-600">Everything we do stays within our secure Randstad Google environment. We are not sharing potentially sensitive client or job information with a public model.</p></div><div><h4 class="google-sans text-xl font-medium">2. Workflow Integration</h4><p class="text-gray-600">With one click, our work can be exported to Docs or drafted in Gmail. This is a huge efficiency gain. For 95% of our recruiting tasks, Gemini is the right choice.</p></div></div></div><div class="flex justify-center"><svg xmlns="http://www.w3.org/2000/svg" class="h-28 w-28 text-blue-500" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="1"><path stroke-linecap="round" stroke-linejoin="round" d="M9 12l2 2 4-4m5.618-4.016A11.955 11.955 0 0112 2.944a11.955 11.955 0 01-8.618 3.04A12 12 0 0012 21.954a12 12 0 008.618-3.04A11.955 11.955 0 0112 2.944z" /></svg></div></div><div class="content-section"><div class="alert-box bg-red-50 border-l-4 border-red-500 p-6 rounded"><h3 class="google-sans text-2xl font-bold text-red-700 mb-2">The Security Imperative: A Non-Negotiable Rule</h3><p class="text-red-900">You must NEVER paste sensitive PII (Personally Identifiable Information) or Client Data into a public, non-approved AI tool. This includes names, emails, phone numbers, full CVs, client names, and project details. Doing so is a major data privacy and security risk. When in doubt, always default to our secure, internal platform.</p></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Lab Activity: "Choose Your Tool" Scenarios</h3><div class="space-y-6"><div class="bg-white p-6 rounded-lg border"><p class="font-medium mb-2">Scenario 1:</p><p class="mb-3">You need to summarize your confidential interview notes about three candidates for a sensitive role with our top client.</p><p class="font-bold text-blue-700">Which tool do you use? (Answer: Gemini)</p></div><div class="bg-white p-6 rounded-lg border"><p class="font-medium mb-2">Scenario 2:</p><p class="mb-3">You want to brainstorm some fun, catchy, generic slogans for a university career fair flyer.</p><p class="font-bold text-blue-700">Which tool do you use? (Answer: Gemini)</p></div><div class="bg-white p-6 rounded-lg border"><p class="font-medium mb-2">Scenario 3:</p><p class="mb-3">You need to find a 30-minute slot for an interview with your candidate, two hiring managers, and one HR partner.</p><p class="font-bold text-blue-700">Which tool do you use? (Answer: A specialized AI Scheduling Tool)</p></div></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Key Takeaways & Next Steps</h3><p class="max-w-2xl mx-auto mb-4">You now have a clear framework for choosing the right AI tool. The golden rule is Security First, Always.
                  Use the right tool to work smarter, faster, and more securely.</p><div class="bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 text-left inline-block"><p><strong class="font-bold">Next Session:</strong> Session 2.3 - Responsible AI in Recruiting & Final Showcase.</p></div></div>
