+++ rpo-training/sessions/2-3.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_2_3 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 2.3: Responsible AI & Final Showcase</h2><p class="mt-4 text-lg text-gray-600">Our Goal: To cover the most important topic of becoming responsible and ethical AI users, and then to celebrate your progress with a final showcase.</p></div><div class="content-section grid md:grid-cols-3 gap-8 items-center"><div class="md:col-span-2"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">What is AI Bias?</h3><p class="mb-4">AI models learn from vast amounts of internet data. The problem is that this data reflects the world as it has been, including all of its historical and unconscious biases. If we're not careful, the AI can simply echo and amplify these past biases.</p><p class="font-semibold">Our job is to be the filter that ensures fairness and equity.</p></div><div class="flex justify-center"><svg xmlns="http://www.w3.org/2000/svg" class="h-24 w-24 text-blue-500" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="1"><path stroke-linecap="round" stroke-linejoin="round" d="M3 6l3 1m0 0l-3 9a5.002 5.002 0 006.001 0M6 7l3 9M6 7l6-2m6 2l3-1m-3 1l-3 9a5.002 5.002 0 006.001 0M18 7l3 9m-3-9l-6-2m0-2v2m0 16V5m0 16H9m3 0h3" /></svg></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Where Can Bias Hide in Recruiting?</h3><div class="grid md:grid-cols-2 gap-8"><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">In Job Descriptions</h4><ul class="list-disc list-inside space-y-2 text-gray-600"><li>Using gendered language (e.g., "rockstar," "ninja," "aggressive salesman").</li><li>Listing overly specific requirements that might exclude diverse talent pools (e.g., a degree from only one specific university).</li></ul></div><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">In CV Summaries</h4><ul class="list-disc list-inside space-y-2 text-gray-600"><li>Over-emphasizing traditional career paths and overlooking non-traditional experience.</li><li>Unintentionally associating certain names with certain skills based on biased training data.</li></ul></div></div><p class="text-center font-semibold mt-6">Your critical eye is the most important tool.</p></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Lab Example: Fixing the Prompt</h3><p class="text-center max-w-2xl mx-auto mb-6">In the lab, we practice identifying and correcting potentially biased AI outputs by rewriting the prompt that created them.</p><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-lg font-medium mb-2">The Biased AI Output:</h4><div class="alert-box bg-red-50 border-l-4 border-red-500 p-4 rounded mb-4"><p class="text-red-900">"We're looking for a digital native and recent graduate who is a true code ninja to join our young, energetic team."</p></div><p class="mb-4"><strong>Potential Bias:</strong> Ageism ("recent graduate," "young"), gendered language ("ninja"), excludes experienced candidates.</p><h4 class="google-sans text-lg font-medium mb-2">The Corrected, Inclusive Prompt:</h4><div class="prompt-box">Write a job description for a software developer role. The ideal candidate should be proficient with modern coding practices and passionate about collaborative teamwork. The tone should be inclusive, professional, and welcoming to candidates of all ages and experience levels.</div></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">The Most Important Principle</h3><div class="inline-block bg-white p-6 rounded-lg border"><p class="text-xl font-semibold">You are the "Human in the Loop."</p><p class="mt-2 max-w-xl mx-auto">AI is a powerful assistant, but it is not the decision-maker. You are the final editor, the final judge, and the final quality check. Your industry expertise and human judgment are what make this partnership work.</p><p class="font-bold text-2xl mt-4 text-blue-700">Never Trust, Always Verify.</p></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Final Showcase: The Prompt-a-Thon!</h3><p class="max-w-2xl mx-auto mb-6">Time to celebrate your skills! Each team will present their single best prompt created during the training in a 2-3 minute pitch. We'll see the goal, the prompt, and the amazing result. After all presentations, we'll use a live poll to award the "People's Choice Award" for the most innovative and effective prompt!</p></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Phase 1 Complete & What's Next</h3><p class="max-w-2xl mx-auto mb-4">Congratulations! You have successfully completed the mandatory AI Foundations training. You are all now AI-enabled recruiters.</p><div class="bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 text-left inline-block"><p><strong class="font-bold">The Journey Continues:</strong> For those who want to go deeper, consider the optional next phases:</p><ul class="list-disc list-inside ml-4"><li><strong>Phase 2:</strong> The AI Champions Track</li><li><strong>Phase 3:</strong> The AI for RPO Strategists Track</li></ul></div></div>
 <div id="footer-placeholder"></div>
