+++ rpo-training/sessions/3-1.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_3_1 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 3.1: Data Insights with Gemini in Sheets</h2><p class="mt-4 text-lg text-gray-600">Welcome to the AI Champions Track! Today, we're taking a huge leap forward. You will learn how to make your data talk back to you, moving from being a user of data to becoming an analyst—all without writing a single complex formula.</p></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">Core Skill 1: AI-Powered Data Cleaning</h3><p class="text-lg text-gray-600">Your New Data Janitor</p></div><div class="mt-6 bg-white border rounded-lg p-6"><p class="mb-6 text-center">Forget manually fixing hundreds of rows. Use natural language to clean your entire dataset in seconds.</p><div class="grid md:grid-cols-2 gap-8"><div><h4 class="google-sans text-xl font-medium mb-2">Example 1: Standardizing Data</h4><p class="mb-3"><strong>The Mess:</strong> A column with "NY," "New York," and "nyc."</p><div class="prompt-box">Standardize the 'Location' column to the format 'City, State'.</div></div><div><h4 class="google-sans text-xl font-medium mb-2">Example 2: Extracting Data</h4><p class="mb-3"><strong>The Mess:</strong> A "Notes" column filled with long sentences.</p><div class="prompt-box">Extract all mentioned technical skills from the 'Notes' column and place them in a new 'Skills' column.</div></div></div></div></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">Core Skill 2: Generating Insights</h3><p class="text-lg text-gray-600">Become a Data Detective</p></div><div class="mt-6 bg-white border rounded-lg p-6"><p class="mb-6 text-center">Once your data is clean, you can start asking it questions.</p><div class="space-y-4"><div class="p-4 bg-gray-50 rounded-lg"><h4 class="google-sans font-medium">1. Asking Direct Questions</h4><p class="text-gray-600">"Which recruiter had the highest number of placements in Q2?"</p></div><div class="p-4 bg-gray-50 rounded-lg"><h4 class="google-sans font-medium">2. Finding Trends</h4><p class="text-gray-600">"Show me the trend of time-to-fill over the last 6 months as a line chart."</p></div><div class="p-4 bg-gray-50 rounded-lg"><h4 class="google-sans font-medium">3. Creating Visuals</h4><p class="text-gray-600">"Create a pie chart showing the distribution of candidates by source."</p></div></div></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Lab Activities</h3><div class="grid md:grid-cols-2 gap-8"><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Lab 1: The Data Janitor (20 Mins)</h4><p><strong>Your Goal:</strong> Clean and structure a messy, real-world-style dataset using only natural language prompts.</p><ul class="list-disc list-inside mt-4 space-y-1"><li>Open the sample Google Sheet provided.</li><li>Use Gemini prompts to clean at least three different columns.</li><li>Create one new column by extracting information from another.</li><li class="font-semibold">Be ready to share your most effective cleaning prompt!</li></ul></div><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Lab 2: The Data Detective (20 Mins)</h4><p><strong>Your Goal:</strong> To answer three specific business questions from your newly cleaned dataset using only natural language prompts.</p><ul class="list-disc list-inside mt-4 space-y-1"><li>What is the average time-to-fill for roles in Texas vs. California?</li><li>Who are the top 3 candidate sources that lead to hires?</li><li class="font-semibold">Create a bar chart showing the number of hires per month for the last 6 months.</li></ul></div></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Key Takeaways & Next Steps</h3><p class="max-w-2xl mx-auto mb-4">Your new superpower is the ability to turn any spreadsheet into an interactive dashboard. The golden rule is that clean data is the foundation for good insights. Use AI to automate the cleaning so you can focus on the analysis.</p><div class="bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 text-left inline-block"><p><strong class="font-bold">Next Session:</strong> Session 3.2 - Building an Intelligent Knowledge Base (No-Code RAG).</p></div></div>
 <div id="footer-placeholder"></div>
