+++ rpo-training/sessions/3-2.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_3_2 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 3.2: Building an Intelligent Knowledge Base (No-Code RAG)</h2><p class="mt-4 text-lg text-gray-600">Let's ask a question that was impossible yesterday: What if you could "talk" to your entire team's history of CVs and interview notes? Today, we're going to build the tool that does exactly that.</p></div><div class="content-section grid md:grid-cols-3 gap-8 items-center"><div class="md:col-span-2"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">What is RAG? (The Simple Explanation)</h3><p class="mb-4">It stands for Retrieval-Augmented Generation, and it sounds complicated, but the concept is simple.</p><div class="quote-box"><p class="text-lg italic">"RAG is the process of giving a brilliant expert (the AI) your private library of documents to read before they answer your question."</p></div><p class="mt-4">This ensures the answers are grounded in your data, not the public internet. It's secure, relevant, and incredibly powerful.</p></div><div class="flex justify-center"><svg xmlns="http://www.w3.org/2000/svg" class="h-28 w-28 text-blue-500" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="1"><path stroke-linecap="round" stroke-linejoin="round" d="M4 6h16M4 12h16M4 18h7" /><path stroke-linecap="round" stroke-linejoin="round" d="M15 18h5a2 2 0 002-2V8a2 2 0 00-2-2h-5l-2 4 2 4zM4 6a2 2 0 012-2h12a2 2 0 012 2v12a2 2 0 01-2-2H6a2 2 0 01-2-2V6z" /></svg></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">The Power of RAG: Meet Your "CV & Candidate Info Bot"</h3><p class="text-center max-w-2xl mx-auto mb-6">This is what our final product will be able to do. Imagine asking a chat bot these questions about your team's private CV database:</p><div class="bg-white p-6 rounded-lg border"><ul class="list-disc list-inside space-y-3 text-gray-700"><li>"Which candidates have experience with both Python and AWS?"</li><li>"Summarize Jane Doe's experience and highlight her top 3 skills."</li><li>"Who is the most experienced project manager for a fintech role based in London?"</li></ul></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Today's Project: Build Your Bot</h3><p class="text-center max-w-2xl mx-auto mb-8">This is a guided, hands-on lab. We will walk through every step together to create your first AI application.</p><div class="grid md:grid-cols-2 gap-8"><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Part 1: The Build (20 Mins)</h4><p>We'll use a no-code tool (like Vertex AI Search) to create a secure "Data Store" and upload a provided set of 20-30 anonymized sample CVs. The system will then "read" and "index" this information.</p></div><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Part 2: The Test (40 Mins)</h4><p>Once indexing is complete, you'll get access to the chat interface for your bot. Your task is to complete the "Scavenger Hunt" to test its capabilities.</p></div></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Lab Task: The Scavenger Hunt</h3><p class="text-center max-w-2xl mx-auto mb-6">Use the chat interface for the bot you just built to find the answers to these questions:</p><div class="bg-white p-6 rounded-lg border"><ol class="list-decimal list-inside space-y-3"><li>Find all candidates with a PMP certification.</li><li>Which candidates mention "data analysis" as a skill?</li><li>Summarize the CV of the candidate named "John Smith".</li><li>Who has more years of experience in management, "Jane Doe" or "Peter Jones"?</li><li class="font-semibold">Which candidate is the best fit for a role requiring experience in both marketing and e-commerce? Explain why.</li></ol></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Key Takeaways & Next Steps</h3><p class="max-w-2xl mx-auto mb-4">Your new superpower is that you have just built one of the most powerful tools in modern AI: a secure application that can reason over your team's collective knowledge. The golden rule is that the power of RAG is its ability to ground AI in trusted, specific data.</p><div class="bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 text-left inline-block"><p><strong class="font-bold">Next Module:</strong> Module 4 - Automating Recruiting Workflows.</p></div></div>
 <div id="footer-placeholder"></div>
