+++ rpo-training/sessions/4-1.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_4_1 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 4.1: Introduction to No-Code Automation</h2><p class="mt-4 text-lg text-gray-600">Our Goal: To build a workflow that completely automates a boring task. Let's connect the dots and build a system that works for you, even when you're not looking.</p></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">The "Finished Product" Demo</h3><p class="text-lg text-gray-600">See Where We're Going</p></div><div class="mt-6 bg-white border rounded-lg p-6"><p class="text-center mb-4">Imagine this hands-free workflow we are building today:</p><ol class="list-decimal list-inside space-y-3 text-lg"><li>An email with candidate feedback arrives in your Gmail and you give it a label.</li><li>You do nothing else.</li><li>A moment later, a new row appears in your tracking spreadsheet, automatically filled with the candidate's name, a one-sentence summary, and a sentiment score.</li></ol></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">How Does Automation Actually Work?</h3><div class="grid md:grid-cols-3 gap-6"><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">APIs</h4><p>Think of an API as a waiter in a restaurant. It's how different apps talk to each other, taking requests and returning data.</p></div><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">Triggers</h4><p>This is the "When this happens..." part of the equation. For us, it's "When a new email gets the label 'Candidate Feedback'..."</p></div><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">Actions</h4><p>This is the "...Then do that" part. For us, it's "...run my script to analyze it and add it to the spreadsheet."</p></div></div></div><div class="content-section grid md:grid-cols-2 gap-8 items-center"><div><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Our Platform: Google Apps Script</h3><p class="mb-4">We will be using Google Apps Script. It's built into your Google account and is perfect for connecting Google products.</p><ul class="list-disc list-inside space-y-2"><li>It uses a simplified version of JavaScript.</li><li class="font-semibold">Don't worry! You do not need to be a coder.</li><li>We will provide you with a well-commented code template. Your job is to understand the logic, not to write code from scratch.</li></ul></div><div class="flex justify-center"><svg xmlns="http://www.w3.org/2000/svg" class="h-28 w-28 text-blue-500" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="1"><path stroke-linecap="round" stroke-linejoin="round" d="M10 20l4-16m4 4l4 4-4 4M6 16l-4-4 4-4" /></svg></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Lab Activity: The End-to-End Workflow (65 Mins)</h3><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Your Goal:</h4><p class="mb-4">To build and test the full, automated workflow from email to spreadsheet.</p><h4 class="google-sans text-xl font-medium mb-2">The Plan of Action:</h4><ol class="list-decimal list-inside space-y-2"><li><strong>The Connection:</strong> Use the provided template to create a script and set up the trigger to read emails from Gmail.</li><li><strong>The AI:</strong> Add the Gemini API call to the script to analyze the email's content.</li><li><strong>The Destination:</strong> Add the final lines of code to populate your Google Sheet.</li><li><strong>The Test:</strong> Send a sample email to yourself, label it, and watch the magic happen!</li></ol></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">The "Magic" Prompt</h3><p class="text-center max-w-2xl mx-auto mb-6">This is the key to making the workflow intelligent. This prompt forces the AI to give us back perfectly structured data that our spreadsheet can understand.</p><div class="prompt-box">From the following email text, extract the candidate's name, provide a one-sentence summary of the feedback, and classify the interviewer's sentiment as Positive, Negative, or Neutral. Return this as a JSON object.</div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Key Takeaways & Next Steps</h3><p class="max-w-2xl mx-auto mb-4">Your new superpower is that you didn't just run a script; you built a system. The golden rule is to identify the most repetitive, copy-paste parts of your job, as these are the perfect candidates for automation.</p><div class="bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 text-left inline-block"><p><strong class="font-bold">Next Module:</strong> Module 5 - Train the Trainer & Project Showcase.</p></div></div>
 <div id="footer-placeholder"></div>
