+++ rpo-training/sessions/5-1.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_5_1 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 5.1: Becoming an AI Champion</h2><p class="mt-4 text-lg text-gray-600">Your journey so far has made you a skilled power user. This session is about the final, most important step: learning how to share that knowledge. You are now the AI leaders for your team. Your role is to empower, guide, and support.</p></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">The Mentor Mindset</h3><p class="text-lg text-gray-600">Coach, Don't Just Tell</p></div><div class="mt-6 bg-white border rounded-lg p-6"><p class="text-center mb-6">The biggest shift you will make is from being the "expert with the answers" to becoming an "enabler of others' success."</p><div class="grid md:grid-cols-2 gap-8"><div><h4 class="google-sans text-xl font-medium mb-2">Telling</h4><p class="italic">"Here, let me just write the prompt for you."</p><p class="mt-2 text-sm text-red-600"><strong>Result:</strong> Solves one problem, but doesn't build skill.</p></div><div><h4 class="google-sans text-xl font-medium mb-2">Coaching</h4><p class="italic">"That's a great question. What have you tried so far?"</p><p class="mt-2 text-sm text-green-600"><strong>Result:</strong> Builds your colleague's confidence and problem-solving ability.</p></div></div><p class="text-center font-bold text-xl mt-8 text-blue-700">Your goal is to make yourself redundant. Teach them how to fish.</p></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">The "Ask, Don't Tell" Model in Action</h3><p class="text-center max-w-2xl mx-auto mb-6">Instead of giving the answer, use these guiding questions to help your colleagues find their own solution.</p><div class="grid md:grid-cols-2 gap-4"><div class="bg-white p-4 rounded-lg border">"What's the ideal output you're looking for?"</div><div class="bg-white p-4 rounded-lg border">"What's the most important piece of context the AI might be missing?"</div><div class="bg-white p-4 rounded-lg border">"If you were to give the AI a role, what would it be?"</div><div class="bg-white p-4 rounded-lg border">"What have you tried so far, and what were the results?"</div></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">How to Lead a Great Prompt Clinic</h3><p class="text-center max-w-2xl mx-auto mb-8">When a colleague is stuck, use this simple 5-step framework to guide a group problem-solving session.</p><div class="space-y-4"><div class="step-box"><p><strong class="google-sans text-lg">1. Set the Stage:</strong> Start by saying, "This is a no-judgment zone. Let's solve this together."</p></div><div class="step-box"><p><strong class="google-sans text-lg">2. Understand the Goal:</strong> Let the person explain what they want to achieve.</p></div><div class="step-box"><p><strong class="google-sans text-lg">3. Open the Floor:</strong> Ask the group, "What ideas does the team have?"</p></div><div class="step-box"><p><strong class="google-sans text-lg">4. Guide, Don't Drive:</strong> Help the user refine their own prompt based on the group's feedback.</p></div><div class="step-box"><p><strong class="google-sans text-lg">5. Summarize the Win:</strong> End with, "Great! The key change that worked here was adding more context."</p></div></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Lab Activities</h3><div class="grid md:grid-cols-2 gap-8"><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Lab 1: Mock Prompt Clinic (20 Mins)</h4><p><strong>Your Goal:</strong> To practice facilitating a problem-solving session in a safe environment.</p><ul class="list-disc list-inside mt-4 space-y-1"><li>Break out into small groups.</li><li>Each person takes a turn leading a 10-minute mini-clinic.</li><li class="font-semibold">Focus on the process
                  of facilitation, not just the technical answer.</li></ul></div><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Lab 2: Brainstorming Responses (15 Mins)</h4><p><strong>Your Goal:</strong> To collectively brainstorm effective and reassuring responses to common concerns like "I don't have time for this."</p><ul class="list-disc list-inside mt-4 space-y-1"><li>Focus on benefits, starting small, and offering support.</li><li class="font-semibold">Example: "I felt that way too at first, but let me show you one trick that saved me 30 minutes yesterday."</li></ul></div></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Key Takeaways & Next Steps</h3><p class="max-w-2xl mx-auto mb-4">Your new superpower is that you are now a coach and a facilitator. Your success is now measured by the success of those you help. Empower, guide, and support.</p><div class="bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 text-left inline-block"><p><strong class="font-bold">The Final Step:</strong> Session 5.2 - The Capstone Project Showcase.</p><p>This is your chance to present the amazing tools you built to leadership and demonstrate the real-world value of AI to the business.</p></div></div>
