+++ rpo-training/sessions/5-2.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_5_2 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 5.2: Capstone Project Showcase</h2><p class="mt-4 text-lg text-gray-600">This is not a training session; it is a formal presentation to leadership. The objective is to demonstrate the tangible business value, ROI, and innovative potential of the tools our AI Champions have built.</p></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">The Keys to a Successful Presentation</h3><p class="text-lg text-gray-600">How to Communicate Your Project's Value</p></div><div class="mt-8 grid md:grid-cols-2 gap-6"><div class="key-point-box"><h4 class="google-sans text-xl font-medium mb-2">1. Speak in Business Terms</h4><p>Translate features into benefits. Instead of "We built a RAG application," say "We built a tool that reduces time-to-find-past-candidates by 90%."</p></div><div class="key-point-box"><h4 class="google-sans text-xl font-medium mb-2">2. Demonstrate Clear ROI</h4><p>Articulate the return on investment through metrics like time saved, improved data quality, or enhanced sourcing capabilities.</p></div><div class="key-point-box"><h4 class="google-sans text-xl font-medium mb-2">3. Tell a Compelling Story</h4><p>Structure your presentation as a narrative: The Problem &rarr; Our Journey &rarr; The Solution &rarr; The Future Impact.</p></div><div class="key-point-box"><h4 class="google-sans text-xl font-medium mb-2">4. Influence Leadership</h4><p>Make a clear and confident case for continued support and adoption of your project.</p></div></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">The Presentation Structure (10 Mins Total)</h3><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-4 text-center">Your Blueprint for a Powerful Pitch</h4><ul class="list-none space-y-4"><li><strong>The Problem (1 min):</strong> "What was the manual process or business challenge we were facing?"</li><li><strong>Our Solution (1 min):</strong> "We built an automated workflow / an intelligent knowledge base to solve this." Briefly explain what it is.</li><li><strong>Live Demo (3-4 mins):</strong> A live, crisp demonstration of the tool in action. Practice this!</li><li><strong>The Impact & ROI (1-2 mins):</strong> "This tool saves an estimated X hours per week," or "This allows us to answer questions that were previously impossible." Quantify the value.</li><li><strong>Q&A with Leadership (3 mins):</strong> Be prepared to answer questions about scalability, cost, and impact.</li></ul></div></div><div class="content-section grid md:grid-cols-2 gap-8 items-start"><div><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Today's Agenda</h3><table class="w-full mt-4"><thead><tr><th>Time (Mins)</th><th>Section</th></tr></thead><tbody><tr><td><strong>0-5</strong></td><td>Opening Remarks by GBS Leadership</td></tr><tr><td><strong>5-75</strong></td><td>AI Champion Project Presentations</td></tr><tr><td><strong>75-85</strong></td><td>Leadership Panel Q&A & Feedback</td></tr><tr><td><strong>85-90</strong></td><td>Closing Remarks & "Graduation"</td></tr></tbody></table></div><div><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Final Checklist for Presenters</h3><ul class="list-none bg-white p-6 rounded-lg border space-y-3"><li>[ ] Is your project finalized and tested?</li><li>[ ] Have you rehearsed your presentation to fit the 7-minute time limit?</li><li>[ ] Is your ROI calculation clear and easy to understand?</li><li>[ ] Are you prepared to answer tough questions from leadership?</li><li>[ ] Are you ready to showcase the incredible work you've done?</li></ul><p class="text-center font-semibold mt-4">Good luck, Champions. This is your moment to shine.</p></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">"Graduation" & The Path Forward</h3><div class="quote-box text-left"><h4 class="google-sans text-xl font-medium mb-2">Congratulations, AI Champions!</h4><ul class="list-disc list-inside space-y-2"><li><strong>Official Recognition:</strong> You are the inaugural graduating class of AI Champions. Your dedication has paved the way for the future of AI in our organization.</li><li><strong>Your Ongoing Role:</strong> You are now the internal consultants, mentors, and leaders who will drive the adoption of these powerful new tools across the team.</li><li><strong>Business Commitment:</strong> Leadership will now discuss how to best support the rollout and scaling of the projects you've seen today.</li></ul></div><p class="mt-6 font-semibold text-lg">Thank you for your vision and hard work.</p></div>
 <div id="footer-placeholder"></div>
