+++ rpo-training/sessions/6-1.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_6_1 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 6.1: Developing an RPO-Specific AI Roadmap</h2><p class="mt-4 text-lg text-gray-600">A Strategic Workshop for RPO Leadership. Our goal is to collaboratively build the strategic map that will guide our team's AI efforts for the next year, ensuring we invest in projects that deliver maximum business value.</p></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">The Strategic Foundation</h3><p class="text-lg text-gray-600">Connecting Capabilities to Business Goals</p></div><div class="mt-6 grid md:grid-cols-2 gap-8"><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">1. Our Team's Proven Capabilities</h4><p>We'll review the successful projects from the Phase 2 "AI Champions" showcase (e.g., the CV Bot, automated workflows). These are the tools we know we can build.</p></div><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">2. Our Department's Business Objectives</h4><p>We'll review our key goals for the year (e.g., reduce time-to-fill by 15%, improve source-of-hire quality, etc.).</p></div></div><div class="mt-8 quote-box"><p class="text-lg text-center font-semibold">The Key Question for Today: "How can we apply our proven AI capabilities to directly accelerate our most important business objectives?"</p></div></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">Workshop Part 1: Opportunity Mapping</h3><p class="text-lg text-gray-600">Brainstorming Across the Recruiting Lifecycle</p></div><div class="mt-6 bg-white border rounded-lg p-6"><p class="mb-4"><strong>The Goal:</strong> To create a comprehensive list of all potential AI projects.</p><p class="mb-4"><strong>The Activity:</strong> Using a digital whiteboard, we will map out the entire RPO lifecycle:</p><p class="text-center font-semibold text-gray-700 bg-gray-100 p-3 rounded-md">Intake &rarr; Sourcing &rarr; Screening &rarr; Submitting &rarr; Scheduling &rarr; Onboarding</p><p class="mt-4">For each stage, we will brainstorm answers to two questions:</p><ol class="list-decimal list-inside mt-2 space-y-2"><li>What are our biggest pain points or bottlenecks in this stage?</li><li>How could an AI tool or automation help solve them?</li></ol></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">The Prioritization Tool: The Impact vs. Effort Matrix</h3><p class="text-center max-w-2xl mx-auto mb-6">Not all ideas are created equal. We will use this simple but powerful matrix to prioritize our list and spark crucial strategic conversations about value versus resources.</p><div class="bg-white p-6 rounded-lg border"><div class="grid grid-cols-2 gap-4"><div class="border p-4 rounded-lg bg-green-50"><h4 class="google-sans font-bold text-lg text-green-800">Quick Wins</h4><p class="text-sm text-green-700">(High Impact / Low Effort)</p><p class="mt-2">Do these now!</p></div><div class="border p-4 rounded-lg bg-blue-50"><h4 class="google-sans font-bold text-lg text-blue-800">Major Projects</h4><p class="text-sm text-blue-700">(High Impact / High Effort)</p><p class="mt-2">Plan for these.</p></div><div class="border p-4 rounded-lg bg-yellow-50"><h4 class="google-sans font-bold text-lg text-yellow-800">Fill-ins</h4><p class="text-sm text-yellow-700">(Low Impact / Low Effort)</p><p class="mt-2">Consider these later.</p></div><div class="border p-4 rounded-lg bg-red-50"><h4 class="google-sans font-bold text-lg text-red-800">Thankless Tasks</h4><p class="text-sm text-red-700">(Low Impact / High Effort)</p><p class="mt-2">Avoid these.</p></div></div></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">From a Map to a Plan</h3><div class="grid md:grid-cols-2 gap-8"><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Workshop Part 2: Prioritizing</h4><p><strong>The Goal:</strong> To create a prioritized visual map of all brainstormed ideas.</p><p class="mt-2">As a leadership team, we will collectively discuss each idea from the Opportunity Mapping exercise and place it onto the Impact/Effort matrix.</p></div><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Workshop Part 3: Building the Roadmap</h4><p><strong>The Goal:</strong> To select our top projects and create a clear, sequenced plan of action.</p><p class="mt-2">We will identify the top 3-5 initiatives from the "Quick Wins" and "Major Projects" quadrants and sequence them on a simple timeline (e.g., Q3, Q4, Next Q1).</p></div></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">Key Takeaways & Next Steps</h3><div class="bg-white p-6 rounded-lg border inline-block text-left"><h4 class="google-sans text-xl font-medium mb-2">What We've Achieved Today:</h4><ul class="list-disc list-inside space-y-2"><li>A shared vision for AI in our department.</li><li>A comprehensive list of potential opportunities.</li><li>A prioritized list of our top 3-5 projects.</li><li>A draft roadmap for the next 6-9 months.</li></ul><p class="font-semibold mt-4">This is our strategic foundation.</p></div><div class="mt-6 bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 text-left inline-block"><p><strong class="font-bold">Next Session:</strong> Session 7.1 - The ROI of AI in Recruiting.</p><p>We will focus on defining how we will measure the success and calculate the ROI for these specific projects.</p></div></div>
 <div id="footer-placeholder"></div>
