+++ rpo-training/sessions/7-1.html
@@ -1,4 +1,4 @@
-<button onclick="navigateTo('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
+<button onclick="NavigationManager.navigate('main-page')" class="google-sans my-6 text-blue-600 font-semibold hover:underline">&larr; Back to Training Overview</button>
             <!-- Content from google_site_training_7_1 -->
             <div class="content-section text-center"><h2 class="google-sans text-4xl font-bold text-gray-800">Session 7.1: The ROI of AI in Recruiting</h2><p class="mt-4 text-lg text-gray-600">A Strategic Workshop for RPO Leadership. Our goal is to build a data-driven, compelling business case for our top-priority project, turning our strategic ideas into funded, measurable projects.</p></div><div class="content-section"><div class="text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-2">Defining Success</h3><p class="text-lg text-gray-600">What Does "Good" Look Like? Moving Beyond Anecdotes</p></div><div class="mt-6 grid md:grid-cols-2 gap-8"><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">Leading Indicators</h4><p>These are the early signals of progress that show we're on the right track.</p><ul class="list-disc list-inside mt-2 text-gray-600"><li>Recruiter time saved per week</li><li>Number of qualified candidates sourced via an AI tool</li></ul></div><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">Lagging Indicators</h4><p>These are the long-term business impact metrics that prove the value of our work.</p><ul class="list-disc list-inside mt-2 text-gray-600"><li>Reduction in average time-to-fill</li><li>Improvement in 6-month new-hire retention</li></ul></div></div><div class="mt-8 quote-box"><p class="text-lg text-center font-semibold">We need a mix of both to tell the full story of our success.</p></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Calculating ROI: The Two Sides of Value</h3><div class="grid md:grid-cols-2 gap-8"><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Hard ROI: The Financial Case</h4><p class="mb-4">This is the quantifiable, bottom-line impact. It's the number that gets budgets approved.</p><p class="font-medium">Formula:</p><div class="formula-box mt-2">(Time Saved) x (# of People) x (Avg. Salary) = $$$ Saved</div></div><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2">Soft ROI: The Strategic Case</h4><p class="mb-4">This is the harder-to-measure but equally important strategic value that drives long-term success.</p><ul class="list-disc list-inside text-gray-600"><li>Improved candidate quality</li><li>Higher hiring manager satisfaction</li><li>Enhanced team morale & retention</li></ul></div></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">The Simple Math of Hard ROI</h3><div class="bg-white p-6 rounded-lg border"><h4 class="google-sans text-xl font-medium mb-2 text-center">A Conservative Calculation Example</h4><p class="text-center mb-4"><strong>The Project:</strong> Automating the summarization of interview feedback.</p><div class="space-y-3"><p><strong>The Assumption:</strong> This saves each of our 20 recruiters an average of 2 hours per week.</p><p><strong>The Calculation:</strong></p><ul class="list-none space-y-2 bg-gray-50 p-4 rounded-md"><li>20 recruiters x 2 hours/week = <strong>40 hours saved per week.</strong></li><li>40 hours/week x 50 weeks/year = <strong>2,000 hours saved per year.</strong></li><li class="font-bold">2,000 hours x [Average Recruiter Hourly Rate] = $X in annual productivity gain.</li></ul><p class="text-center font-semibold mt-4">This is a powerful, conservative number to anchor your business case.</p></div></div></div><div class="content-section"><h3 class="google-sans text-2xl font-bold text-center text-blue-600 mb-6">Workshop Activities</h3><div class="space-y-8"><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">Part 1: KPI Mapping (30 Mins)</h4><p>Select your #1 project from the roadmap and answer: "If this project is a massive success one year from now, what 3-4 numbers on our dashboard will have changed?" Define at least two "hard" and one "soft" metric.</p></div><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">Part 2: Building the Business Case (30 Mins)</h4><p>Draft a concise, one-page business case for your chosen project using the provided template. The template will include: Problem Statement, Proposed AI Solution, Key Metrics (KPIs), Estimated ROI, and Required Resources.</p></div><div class="step-box"><h4 class="google-sans text-xl font-medium mb-2">Part 3: Peer Review & Final Refinement (15 Mins)</h4><p>In pairs, present your one-page business cases to each other. Your partner's job is to act as a skeptical leader: ask tough questions, point out weaknesses, and help you strengthen your argument.</p></div></div></div><div class="content-section text-center"><h3 class="google-sans text-2xl font-bold text-blue-600 mb-4">From Vision to Action Plan</h3><p class="max-w-2xl mx-auto mb-4">After today, you will have a clear set of KPIs, a defensible method for calculating ROI, and a drafted, peer-reviewed business case for your top priority. You are now fully prepared to present your vision to GBS leadership.</p><div class="bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 text-left inline-block"><p><strong class="font-bold">The Final Step:</strong> The Capstone Presentation.</p><p>You will now formalize these plans into a final presentation to secure the resources needed to make our AI strategy a reality.</p></div></div>
 <div id="footer-placeholder"></div>
